refactor: extract utils to separate module inside `sql`
