+++ crates/store/src/db/sql/queries/select_merged_deltas.sql
